+++ zimui/src/types/Channel.ts
@@ -7,10 +7,10 @@ export interface Channel {
   profilePath?: string
   bannerPath?: string
   joinedDate: string
-  mainPlaylist?: string
-  userLongUploadsPlaylist?:string
-  userShortUploadsPlaylist?:string
-  userLivesPlaylist?:string
+  firstPlaylist?: string
+  userLongUploadsPlaylist?: string
+  userShortUploadsPlaylist?: string
+  userLivesPlaylist?: string
   playlistCount: number
 }